pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brainstormer"
version = "2026.12.0"
description = "The AI Development Operating System. Drop in. Init. Ship."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Patrick McGuire" },
]
keywords = ["ai", "development", "cli", "scaffolding", "obsidian", "agents"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/brainstormer-dev/brainstormer"
Issues = "https://github.com/brainstormer-dev/brainstormer/issues"
Changelog = "https://github.com/brainstormer-dev/brainstormer/blob/main/CHANGELOG.md"

[project.scripts]
brainstormer = "cli.brainstormer:main"

[tool.setuptools.packages.find]
include = ["cli*"]

[tool.setuptools.package-data]
"*" = [
    "*.md",
    "*.json",
    "*.yaml",
    "*.yml",
    "*.sh",
    "*.bat",
]
